moved constant
